% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqr-ashw.R
\name{ashw}
\alias{ashw}
\alias{dashw}
\alias{pashw}
\alias{qashw}
\alias{rashw}
\title{The arcsecant hyperbolic Weibull distribution}
\usage{
dashw(x, mu, theta, tau = 0.5, log = FALSE)

pashw(q, mu, theta, tau = 0.5, lower.tail = TRUE, log.p = FALSE)

qashw(p, mu, theta, tau = 0.5, lower.tail = TRUE, log.p = FALSE)

rashw(n, mu, theta, tau = 0.5)
}
\arguments{
\item{x, q}{vector of positive quantiles.}

\item{mu}{location parameter indicating the \eqn{\tau}-th quantile, \eqn{\tau \in (0, 1)}.}

\item{theta}{shape parameter.}

\item{tau}{the parameter to specify which quantile use in the parametrization.}

\item{log, log.p}{logical; If TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; If TRUE, (default), \eqn{P(X \leq x)} are returned, otherwise \eqn{P(X > x)}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}
}
\value{
\code{dashw} gives the density, \code{pashw} gives the distribution function,
\code{qashw} gives the quantile function and \code{rashw} generates random deviates.

Invalid arguments will return an error message.
}
\description{
Density function, distribution function, quantile function and random number generation function
for the arcsecant hyperbolic Weibull distribution reparametrized in terms of the \eqn{\tau}-th quantile, \eqn{\tau \in (0, 1)}.
}
\details{
Probability density function
\deqn{f(y;\alpha, \theta)=\frac{\alpha \theta}{y\sqrt{1-y^2}} arcsech(y)^{\theta-1}\exp\left [ -\alpha arcsech(y)^\theta \right ]}

Cumulative distribution function
\deqn{F(y;\alpha, \theta)=\exp\left [ -\alpha arcsech(y)^\theta \right ]}

Quantile function
\deqn{Q(\tau;\alpha, \theta)= sech\left \{ \left [ -\alpha^{-1} \log(\tau)\right ]^{\frac{1}{\theta}} \right \}}

Reparameterization
\deqn{\alpha = g^{-1}(\mu) = -\frac{\log(\tau)}{arcsech(\mu)^\theta}}

where \eqn{\theta >0} is the shape parameter and \eqn{arcsech(y)= \log\left[\left( 1+\sqrt{1-y^2} \right)/y \right]}.
}
\examples{
set.seed(6969)
x <- rashw(n = 1000, mu = 0.5, theta = 2.5, tau = 0.5)
R <- range(x)
S <- seq(from = R[1], to = R[2], by = 0.01)
hist(x, prob = TRUE, main = 'arcsecant hyperbolic Weibull')
lines(S, dashw(x = S, mu = 0.5, theta = 0.5, tau = 0.5), col = 2)
plot(ecdf(x))
lines(S, pashw(q = S, mu = 0.5, theta = 0.75, tau = 0.5), col = 2)
plot(quantile(x, probs = S), type = "l")
lines(qashw(p = S, mu = 0.5, theta = 0.75, tau = 0.5), col = 2)
}
\references{
Korkmaz, M. C., Chesneau, C. and Korkmaz, Z. S., (2021). A new alternative quantile regression model for the bounded response with educational measurements applications of OECD countries. \emph{Journal of Applied Statistics}, 1--25.
}
\author{
Josmar Mazucheli

Bruna Alves
}
