% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodyfat.R, R/datasets.R
\docType{data}
\name{bodyfat}
\alias{bodyfat}
\title{Percentage of body fat data set}
\format{
A data-frame with 298 observations and 13 columns:

\itemize{
\item \code{id}: Subject id.
\item \code{arms}: Arms fat percentage.
\item \code{legs}: Legs fat percentage.
\item \code{body}: Body fat percentage.
\item \code{android}: Android fat percentage.
\item \code{gynecoid}: Ginecoid fat percentage.
\item \code{bmi}: Body mass index - 24.71577.
\item \code{age}: Age - 46.00.
\item \code{sex}: 1 for female, 2 for male.
\item \code{ipaq}: 0 for ipaq = sedentary, 1 for ipaq = insufficiently active and 2 for ipaq = active.
\item \code{sex2}: 0 for female, 1 for male.
\item \code{ipa1}: 0 for ipaq = 0, 1 for ipaq = 1.
\item \code{ipa2}: 0 for ipaq = 0, 1 for ipaq = 2.
}

A \code{\link{data.frame}} with 298 observations and 9 columns:

\itemize{
\item \code{arms}: Arms fat percentage.
\item \code{legs}: Legs fat percentage.
\item \code{body}: Body fat percentage.
\item \code{android}: Android fat percentage.
\item \code{gynecoid}: Ginecoid fat percentage.
\item \code{bmi}: Body mass index - 24.71577.
\item \code{age}: Age - 46.00.
\item \code{sex}: Sex of individual. Female or male.
\item \code{ipaq}: Factor variable indicating the sedentary, insufficiently active or active.
}
}
\source{
\url{http://www.leg.ufpr.br/doku.php/publications:papercompanions:multquasibeta}

\url{http://www.leg.ufpr.br/doku.php/publications:papercompanions:multquasibeta}
}
\usage{
data(bodyfat, package = "unitquantreg")

data(bodyfat, package = "unitquantreg")
}
\description{
The body fat percentage of individuals assisted in a public hospital in Curitiba, Paraná, Brazil.

The body fat percentage of individuals assisted in a public hospital in Curitiba, Paraná, Brasil.
}
\references{
Petterle, R. R., Bonat, W. H., Scarpin, C. T., Jonasson, T., and Borba, V. Z. C., (2020). Multivariate quasi-beta regression models for continuous bounded data. \emph{The International Journal of Biostatistics}, 1--15, (preprint).

Mazucheli, J., Leiva, V., Alves, B., and Menezes A. F. B., (2021). A new quantile regression for modeling bounded data under a unit Birnbaum-Saunders distribution with applications in medicine and politics. \emph{Symmetry}, \bold{13}(4) 1--21.

Petterle, R. R., Bonat, W. H., Scarpin, C. T., Jonasson, T., and Borba, V. Z. C., (2020). Multivariate quasi-beta regression models for continuous bounded data. \emph{The International Journal of Biostatistics}, 1--15, (preprint).

Mazucheli, J., Leiva, V., Alves, B., and Menezes A. F. B., (2021). A new quantile regression for modeling bounded data under a unit Birnbaum-Saunders distribution with applications in medicine and politics. \emph{Symmetry}, \bold{13}(4) 1--21.
}
\author{
Josmar Mazucheli \email{jmazucheli@gmail.com}

Bruna Alves \email{pg402900@uem.br}

André Felipe Menezes \email{andrefelipemaringa@gmail.com}

Josmar Mazucheli \email{jmazucheli@gmail.com}

Bruna Alves \email{pg402900@uem.br}
}
\keyword{datasets}
