% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.unitquantreg}
\alias{predict.unitquantreg}
\title{Prediction method for \code{unitquantreg} class}
\usage{
\method{predict}{unitquantreg}(
  object,
  newdata,
  type = c("link", "quantile", "shape", "terms"),
  interval = c("none", "confidence"),
  level = 0.95,
  se.fit = FALSE,
  ...
)
}
\arguments{
\item{object}{fitted model object of class \code{\link{unitquantreg}}.}

\item{newdata}{optionally, a data frame in which to look for variables with
which to predict. If omitted, the original observations are used.}

\item{type}{character indicating type of predictions. The options are
\code{link}, \code{quantile}, \code{shape} and \code{terms}.}

\item{interval}{type of interval desired. The options are \code{none} and
\code{confidence}. The "\code{terms}" option returns a matrix giving
the fitted values of each term in the model formula on the linear predictor
scale.}

\item{level}{coverage probability for the confidence intervals. Default is
\code{0.95}.}

\item{se.fit}{logical. If \code{TRUE} return the asymptotic standard errors.}

\item{...}{currently not used.}
}
\description{
Extract various types of predictions from unit quantile regression models.
}
\author{
André F. B. Menezes
}
