% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{water}
\alias{water}
\title{Access to piped water supply data set}
\format{
\code{\link{data.frame}} with 3457 observations and 5 columns:

\itemize{
\item \code{phpws}: the proportion of households with piped water supply.
\item \code{mhdi}: municipal human development index.
\item \code{incpc}: per capita income.
\item \code{region}: 0 for Southeast, 1 for Northeast.
\item \code{pop}: total population.
}
}
\usage{
data(water, package = "unitquantreg")
}
\description{
The access of people in households with piped water supply in the cities of Brazil from
the Southeast and Northeast regions. Information obtained during the census of 2010.
}
\references{
Mazucheli, J., Menezes, A. F. B., Fernandes, L. B., Oliveira, R. P. and Ghitany, M. E., (2020).
The unit-Weibull distribution as an alternative to the Kumaraswamy distribution for the
modeling of quantiles conditional on covariates. \emph{Jounal of Applied Statistics}, \bold{47}(6), 954--974.
}
\author{
André F. B. Menezes
}
\keyword{datasets}
